+++ mvc/src/main/java/com/interface21/webmvc/servlet/BeanFactory.java
@@ -1,5 +1,6 @@
 package com.interface21.webmvc.servlet;
 
+import com.interface21.webmvc.servlet.exception.UncheckedServletException;
 import java.lang.reflect.Constructor;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
