+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServlet.java
@@ -1,5 +1,7 @@
 package com.interface21.webmvc.servlet;
 
+import com.interface21.webmvc.servlet.exception.NotFoundAdapterException;
+import com.interface21.webmvc.servlet.exception.NotFoundHandlerException;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
