+++ mvc/src/main/java/com/interface21/webmvc/servlet/exception/NotFoundAdapterException.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.exception;
 
 public class NotFoundAdapterException extends RuntimeException {
